refactor(ui): Move organization editing and deleting
Move organization edit and delete functions to "Settings" menu option, 
which will only be shown for superusers and organization admins. Require 
manual text confirmation for deleting an organization.
